Updates from feedback
Explicitly discuss public collections and collections that are not
associated with either a user or a group.  Add a public flag to the
ACL and short-circuit access to public collections before making a
call to the identity management system.

Add discussion of using GIDs instead of groups and future work to
make that change.

Be clearer about who can modify the ACL.